Move SplashActivity to splash package
