+++ app/src/androidTestKiwix/java/org/kiwix/kiwixmobile/tests/DownloadTest.java
@@ -19,48 +19,46 @@
 package org.kiwix.kiwixmobile.tests;
 
 
-import android.Manifest;
-import android.support.test.espresso.Espresso;
-import android.support.test.espresso.IdlingPolicies;
-import android.support.test.espresso.ViewInteraction;
-import android.support.test.rule.ActivityTestRule;
-import android.support.test.runner.AndroidJUnit4;
-import android.support.test.rule.GrantPermissionRule;
-import android.test.suitebuilder.annotation.LargeTest;
-import android.util.Log;
+ import android.Manifest;
+ import android.support.test.espresso.Espresso;
+ import android.support.test.espresso.IdlingPolicies;
+ import android.support.test.espresso.ViewInteraction;
+ import android.support.test.rule.ActivityTestRule;
+ import android.support.test.rule.GrantPermissionRule;
+ import android.support.test.runner.AndroidJUnit4;
+ import android.test.suitebuilder.annotation.LargeTest;
+ import android.util.Log;
 
-import com.schibsted.spain.barista.interaction.BaristaSleepInteractions;
+ import com.schibsted.spain.barista.interaction.BaristaSleepInteractions;
 
-import org.junit.After;
-import org.junit.Before;
-import org.junit.BeforeClass;
-import org.junit.Rule;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.kiwix.kiwixmobile.R;
-import org.kiwix.kiwixmobile.utils.KiwixIdlingResource;
-import org.kiwix.kiwixmobile.utils.SplashActivity;
+ import org.junit.After;
+ import org.junit.Before;
+ import org.junit.BeforeClass;
+ import org.junit.Rule;
+ import org.junit.Test;
+ import org.junit.runner.RunWith;
+ import org.kiwix.kiwixmobile.R;
+ import org.kiwix.kiwixmobile.splash.SplashActivity;
+ import org.kiwix.kiwixmobile.utils.KiwixIdlingResource;
 
-import java.util.concurrent.TimeUnit;
+ import java.util.concurrent.TimeUnit;
 
-import static android.support.test.espresso.Espresso.onData;
-import static android.support.test.espresso.Espresso.onView;
-import static android.support.test.espresso.action.ViewActions.click;
-import static android.support.test.espresso.matcher.ViewMatchers.isDisplayed;
-import static android.support.test.espresso.matcher.ViewMatchers.withId;
-import static android.support.test.espresso.matcher.ViewMatchers.withParent;
-import static com.schibsted.spain.barista.assertion.BaristaVisibilityAssertions.assertDisplayed;
-import static com.schibsted.spain.barista.interaction.BaristaClickInteractions.clickOn;
-import static com.schibsted.spain.barista.interaction.BaristaDialogInteractions.clickDialogPositiveButton;
-import static com.schibsted.spain.barista.interaction.BaristaSwipeRefreshInteractions.refresh;
-import static junit.framework.Assert.fail;
-import static org.hamcrest.Matchers.allOf;
-import static org.kiwix.kiwixmobile.testutils.TestUtils.TEST_PAUSE_MS;
-import static org.kiwix.kiwixmobile.testutils.TestUtils.allowPermissionsIfNeeded;
-import static org.kiwix.kiwixmobile.testutils.TestUtils.captureAndSaveScreenshot;
-import static org.kiwix.kiwixmobile.testutils.TestUtils.withContent;
-import static org.kiwix.kiwixmobile.utils.StandardActions.deleteZimIfExists;
-import static org.kiwix.kiwixmobile.utils.StandardActions.enterHelp;
+ import static android.support.test.espresso.Espresso.onData;
+ import static android.support.test.espresso.Espresso.onView;
+ import static android.support.test.espresso.matcher.ViewMatchers.isDisplayed;
+ import static android.support.test.espresso.matcher.ViewMatchers.withId;
+ import static android.support.test.espresso.matcher.ViewMatchers.withParent;
+ import static com.schibsted.spain.barista.assertion.BaristaVisibilityAssertions.assertDisplayed;
+ import static com.schibsted.spain.barista.interaction.BaristaClickInteractions.clickOn;
+ import static com.schibsted.spain.barista.interaction.BaristaSwipeRefreshInteractions.refresh;
+ import static junit.framework.Assert.fail;
+ import static org.hamcrest.Matchers.allOf;
+ import static org.kiwix.kiwixmobile.testutils.TestUtils.TEST_PAUSE_MS;
+ import static org.kiwix.kiwixmobile.testutils.TestUtils.allowPermissionsIfNeeded;
+ import static org.kiwix.kiwixmobile.testutils.TestUtils.captureAndSaveScreenshot;
+ import static org.kiwix.kiwixmobile.testutils.TestUtils.withContent;
+ import static org.kiwix.kiwixmobile.utils.StandardActions.deleteZimIfExists;
+ import static org.kiwix.kiwixmobile.utils.StandardActions.enterHelp;
 
 @LargeTest
 @RunWith(AndroidJUnit4.class)
