+++ app/src/main/java/org/kiwix/kiwixmobile/error/ErrorActivity.java
@@ -15,7 +15,7 @@ import org.kiwix.kiwixmobile.base.BaseActivity;
 import org.kiwix.kiwixmobile.data.ZimContentProvider;
 import org.kiwix.kiwixmobile.data.local.dao.BookDao;
 import org.kiwix.kiwixmobile.library.entity.LibraryNetworkEntity;
-import org.kiwix.kiwixmobile.utils.SplashActivity;
+import org.kiwix.kiwixmobile.splash.SplashActivity;
 
 import java.io.File;
 import java.util.ArrayList;
